feat: moved session class to session package
Just moved something in different packages about sessions.
