+++ src/main/kotlin/Application.kt
@@ -4,12 +4,12 @@ import com.cessup.data.exceptions.AuthorizationException
 import com.cessup.data.services.JwtProvider
 import com.cessup.data.services.userRoutes
 import com.cessup.di.AppModule
-import com.cessup.domain.usecases.AuthenticateUseCase
-import com.cessup.domain.usecases.DeleteUserUseCase
-import com.cessup.domain.usecases.GetUserUseCase
-import com.cessup.domain.usecases.RegisterUserUseCase
-import com.cessup.domain.usecases.ResetPasswordUseCase
-import com.cessup.domain.usecases.UpdateUserDetailsUseCase
+import com.cessup.domain.usecases.session.AuthenticateUseCase
+import com.cessup.domain.usecases.session.DeleteUserUseCase
+import com.cessup.domain.usecases.session.GetUserUseCase
+import com.cessup.domain.usecases.session.RegisterUserUseCase
+import com.cessup.domain.usecases.session.ResetPasswordUseCase
+import com.cessup.domain.usecases.session.UpdateUserDetailsUseCase
 import com.google.inject.Guice
 import io.ktor.http.HttpStatusCode
 import io.ktor.serialization.gson.gson
