+++ src/main/kotlin/domain/usecases/session/GetUserUseCase.kt
@@ -1,4 +1,4 @@
-package com.cessup.domain.usecases
+package com.cessup.domain.usecases.session
 
 import com.cessup.domain.repositories.UserRepository
 import com.google.inject.Inject
