+++ src/main/kotlin/domain/usecases/session/ResetPasswordUseCase.kt
@@ -1,4 +1,4 @@
-package com.cessup.domain.usecases
+package com.cessup.domain.usecases.session
 
 import com.cessup.data.services.Encrypt
 import com.cessup.domain.repositories.UserRepository
