+++ src/test/kotlin/com/cessup/routes/UserRoutesTest.kt
@@ -14,13 +14,13 @@ import com.cessup.data.services.JwtProvider
 import com.cessup.data.services.RegisterUserDetailsRequest
 import com.cessup.data.services.RegisterUserRequest
 import com.cessup.data.services.userRoutes
-import com.cessup.domain.models.UserDetails
-import com.cessup.domain.usecases.AuthenticateUseCase
-import com.cessup.domain.usecases.DeleteUserUseCase
-import com.cessup.domain.usecases.GetUserUseCase
-import com.cessup.domain.usecases.RegisterUserUseCase
-import com.cessup.domain.usecases.ResetPasswordUseCase
-import com.cessup.domain.usecases.UpdateUserDetailsUseCase
+import com.cessup.domain.models.session.UserDetails
+import com.cessup.domain.usecases.session.AuthenticateUseCase
+import com.cessup.domain.usecases.session.DeleteUserUseCase
+import com.cessup.domain.usecases.session.GetUserUseCase
+import com.cessup.domain.usecases.session.RegisterUserUseCase
+import com.cessup.domain.usecases.session.ResetPasswordUseCase
+import com.cessup.domain.usecases.session.UpdateUserDetailsUseCase
 import domain.models.User
 import io.ktor.client.request.post
 import io.ktor.client.request.setBody
